Fix the push method
We're trying to move the `self.head` field out to `next`, but Rust
doesn't want us to do that. This would leave `self` only partially
initialized when we end the borrow and "give it back" to its rightful
owner. That's the one thing you can't do with an `&mut`

Instead, for now we use `mem::replace` to steal a value out of a borrow
by replacing it with another value.

Here we replace `self.head` temporarily with `Link::Empty` before
replacing it with the new head of the list. I'm not gonna lie: this is a
pretty unfortunate thing to have to do.