Fix push method
We want to:

* Check if the list is empty.
* If it's empty, just return None
* If it's not empty
  * remove the head of the list
  * remove its elem
  * replace the list's head with its next
  * return Some(elem)

The key insight is we want to remove things, which means we want to get
the head of the list by value. We certainly can't do that through the
shared reference we get through `&self.head`. We also "only" have a
mutable reference to `self`, so the only way we can move stuff is to
replace it. Looks like we're doing the Empty dance again!