Fix references
Remove another layer of indirection. `&**` looks really janky and it's
starting to look like C++.

Normally Rust isvery good at doing this kind of conversion implicitly,
through a process called deref coercion, where basically it can insert
*'s throughout your code to make it type-check. It can do this because
we have the borrow checker to ensure we never mess up pointers!

But in this case the closure in conjunction with the fact that we have
an `Option<&T>` instead of `&T` is a bit too complicated for it to work
out, so we need to do this for it.