Show turbofish alternative solution
The turbofish, `::<>`, lets us tell the compiler what we think the types
of those generics should be. In this case `::<&Node<T>, _>` says "it
should return a `&Node<T>`, and I don't know/care about that other
type".

This in turn lets the compiler know that `&node` should have deref
coercion applied to it, so we don't need to manually apply all those
*'s!