Add persistent immutable version
`new` creates an empty List.

`append` takes a List and an element and returns a List. Because we
don't change anything, we `clone` the `head` of the old List.

`tail` is the inverse of `append`. It takes a List and returns the whole
list with the first element removed. All it does is clone the second
element in the List (if it exists). This uses `and_then` which is the
Rust version of `flat_map` to remove a level of Option indirection.

`head` is the same as `peek` from the previous List implementation.

We also add some tests to check it behaves how we expect.