Try to add peek_front
This is exactly the same logic as our singly-linked version. The problem
is `RefCell`s.

  error[E0515]: cannot return value referencing temporary value
    --> src/fourth.rs:80:13
     |
  80 |             &node.borrow().elem
     |             ^-------------^^^^^
     |             ||
     |             |temporary value created here
     |             returns a value referencing data owned by the current function

`Ref` and `RefMut` implement `Deref` and `DerefMut` respectively. So for
most intents and purposes they behave exactly like `&T` and `&mut T`.
However, because of how those traits work, the reference that's returned
is connected to the lifetime of the `Ref`, and not the actual `RefCell`.
This means that the `Ref` has to be sitting around as long as we keep
the reference around.

This is in fact necessary for correctness. When a `Ref` gets dropped, it
tells the `RefCell` that it's not borrowed anymore. So if we did manage
to hold onto our reference longer than the `Ref` existed, we could get a
`RefMut` while a reference was kicking around and totally break Rust's
type system in half.

So where does that leave us? We only want to return a reference, but we
need to keep this `Ref` thing around. But as soon as we return the
eference from peek, the function is over and the `Ref` goes out of
scope.

😖

As far as I know, we're actually totally dead in the water here. You
can't totally encapsulate the use of `RefCell`s like that.